Refactor: Update dependencies and simplify RunCommand constructor
