+++ .sdkmanrc
@@ -1,2 +1,2 @@
-java=24.0.1-graalce
+java=25.0.1-open
 maven=3.9.9
