+++ churrera-cli/src/main/java/info/jab/churrera/cli/di/ServiceModule.java
@@ -5,9 +5,9 @@ import info.jab.churrera.cli.service.*;
 import info.jab.churrera.cli.service.handler.ChildWorkflowHandler;
 import info.jab.churrera.cli.service.handler.ParallelWorkflowHandler;
 import info.jab.churrera.cli.service.handler.SequenceWorkflowHandler;
-import info.jab.churrera.util.CursorAgentGeneralEndpoints;
-import info.jab.churrera.util.CursorAgentInformation;
-import info.jab.churrera.util.CursorAgentManagement;
+import info.jab.cursor.client.CursorAgentGeneralEndpoints;
+import info.jab.cursor.client.CursorAgentInformation;
+import info.jab.cursor.client.CursorAgentManagement;
 import info.jab.churrera.util.PmlConverter;
 import info.jab.churrera.workflow.WorkflowParser;
 import dagger.Module;
