+++ churrera-cli/src/main/java/info/jab/churrera/cli/service/AgentLauncher.java
@@ -8,6 +8,7 @@ import info.jab.churrera.workflow.PromptInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.inject.Inject;
 import java.time.LocalDateTime;
 
 /**
